+++ src/test/suite/testFiles/nonTranslatedFile.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+const MyComponent = () => {
+  return <div>Test</div>;
+};